Update README.md, added links
Also renamed "About" to "Project Overview" because that's more or less what it became as I was writing it. It was supposed to be a brief description and not a full project overview, but I have a tendency to write in lengthy detail.

This commit's description probably highlights that further.