Update kernel.cpp to play better with loader.s when linked
I have no idea why the main function accepts two arguments here. The magic number should be contained in the multiboot structure already.

In WYOOS, he suspects this is for error handling. It probably has something to do with the bootloader. I'm making a note here to come back to this and figure it out later. It's the one part of the process that I don't understand.

It is interesting to me that the kernel is essentially just the first program that loads into RAM and is never taken out. That makes so much sense and is so simple. I thought the kernel would be a lot more mysterious, but I'm glad it's somewhat straightforward.

I'm already learning a lot from this project.