Update kernel.cpp to work without a linked library
Should have done this earlier, probably.

Honestly, the faux-printf function that I have here is not ideal for several reasons. There's less control over color, for instance. It's written a bit more haphazardly than the one on OSDev, for second.

Also, in general, I'm not as big of a fan of C++. I have coded in C++ in the past, so I'm familiar with the language, but I'm not sure that I want to commit to it for this entire project.

I'll come back and rewrite everything later if that's the case. For now, I'm staying aligned with WYOOS. I can edit everything after I've bundled it with LFS.