Update kernel.cpp, context with comment
This is speculation on my part, but it would make sense. When we accept arguments in C++, space is defined for them before the function. This probably keeps the magicnumber in an accessible place in kernel memory.

If that is the case, then I understand parts of this code slightly more than the guy who wrote it, which I think would be neat.