Update linker.ld, added spaces to ends of colons
I know I said that whitespace is ignored in ld files, for the most part, but I'm thinking that the spaces around the colons might be important. It's the only thing I can think of right now. If this doesn't fix the syntax error at line 9, I'm going to have to take a break and come back to it.